# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/attendance_sessions:
    get:
      tags:
        - AttendanceSessions
      summary: List AttendanceSessions
      description: Get a paginated list of AttendanceSessions with optional filtering and sorting
      operationId: listAttendanceSessions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceSessionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AttendanceSessions
      summary: Create AttendanceSession
      description: Create a new AttendanceSession
      operationId: createAttendanceSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttendanceSessionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/kiosk_pins:
    get:
      tags:
        - KioskPins
      summary: List KioskPins
      description: Get a paginated list of KioskPins with optional filtering and sorting
      operationId: listKioskPins
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KioskPinList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - KioskPins
      summary: Create KioskPin
      description: Create a new KioskPin
      operationId: createKioskPin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKioskPinRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KioskPin'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    AttendanceSession:
      type: object
      required:
        - id
        - company_id
        - employee_id
        - date
        - check_in
        - source
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        check_in:
          type: string
          format: date-time
        check_out:
          nullable: true
          type: string
          format: date-time
        source:
          type: string
        correction_reason:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateAttendanceSessionRequest:
      type: object
      required:
        - company_id
        - employee_id
        - date
        - check_in
        - source
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        check_in:
          type: string
          format: date-time
        check_out:
          nullable: true
          type: string
          format: date-time
        source:
          type: string
        correction_reason:
          nullable: true
          type: string
        metadata:
          type: object
...
    AttendanceSessionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceSession'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    KioskPin:
      type: object
      required:
        - id
        - company_id
        - employee_id
        - badge_code
        - pin_hash
        - failed_attempts
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        badge_code:
          type: string
        pin_hash:
          type: string
        failed_attempts:
          type: integer
          format: int32
        locked_until:
          nullable: true
          type: string
          format: date-time
        last_attempt_at:
          nullable: true
          type: string
          format: date-time
        expires_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateKioskPinRequest:
      type: object
      required:
        - company_id
        - employee_id
        - badge_code
        - pin_hash
        - failed_attempts
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        badge_code:
          type: string
        pin_hash:
          type: string
        failed_attempts:
          type: integer
          format: int32
        locked_until:
          nullable: true
          type: string
          format: date-time
        last_attempt_at:
          nullable: true
          type: string
          format: date-time
        expires_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    KioskPinList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/KioskPin'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []