# src/presentation/http/kiosk_pin_handler.rs
//! KioskPin REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{KioskPinService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateKioskPinDto, UpdateKioskPinDto, PatchKioskPinDto, KioskPinResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum KioskPinError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for KioskPinError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for KioskPinError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "KIOSKPIN_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "KIOSKPIN_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "KIOSKPIN_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "KIOSKPIN_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for KioskPin.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /kiosk_pins | List with pagination |
/// | POST | /kiosk_pins | Create new |
/// | GET | /kiosk_pins/:id | Get by ID |
/// | PUT | /kiosk_pins/:id | Full update |
/// | PATCH | /kiosk_pins/:id | Partial update |
/// | DELETE | /kiosk_pins/:id | Soft delete |
/// | POST | /kiosk_pins/bulk | Bulk create |
/// | POST | /kiosk_pins/upsert | Upsert |
/// | GET | /kiosk_pins/trash | List deleted |
/// | POST | /kiosk_pins/:id/restore | Restore |
/// | DELETE | /kiosk_pins/empty | Empty trash |
/// | GET | /kiosk_pins/:id/deleted | Get deleted by ID |
/// | DELETE | /kiosk_pins/trash/:id | Permanent delete from trash |
/// | GET | /kiosk_pins/count | Count active entities |
/// | GET | /kiosk_pins/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(KioskPinService::with_repository(repository));
/// let router = create_kiosk_pin_routes(service);
/// ```
pub fn create_kiosk_pin_routes(service: Arc<KioskPinService>) -> Router {
    BackboneCrudHandler::<KioskPinService, KioskPin, CreateKioskPinDto, UpdateKioskPinDto, KioskPinResponseDto>::routes(
        service,
        "/kiosk_pins",
    )
}

/// Create Axum router with only the read (GET) endpoints for KioskPin.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_kiosk_pin_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_kiosk_pin_read_routes(service: Arc<KioskPinService>) -> Router {
    BackboneCrudHandler::<KioskPinService, KioskPin, CreateKioskPinDto, UpdateKioskPinDto, KioskPinResponseDto>::read_routes(
        service,
        "/kiosk_pins",
    )
}

/// Create Axum router with only the write (mutation) endpoints for KioskPin.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_kiosk_pin_write_routes(service: Arc<KioskPinService>) -> Router {
    BackboneCrudHandler::<KioskPinService, KioskPin, CreateKioskPinDto, UpdateKioskPinDto, KioskPinResponseDto>::write_routes(
        service,
        "/kiosk_pins",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_kiosk_pin_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<KioskPinService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_kiosk_pin_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}